+++ connect/src/state.rs
@@ -87,7 +87,7 @@ pub struct ConnectConfig {
     pub initial_volume: u16,
     /// Disables the option to control the volume remotely (default: false)
     pub disable_volume: bool,
-    /// The steps in which the volume is incremented (default: 1024)
+    /// Number of incremental steps (default: 64)
     pub volume_steps: u16,
 }
 
@@ -99,7 +99,7 @@ impl Default for ConnectConfig {
             is_group: false,
             initial_volume: u16::MAX / 2,
             disable_volume: false,
-            volume_steps: 1024,
+            volume_steps: 64,
         }
     }
 }
@@ -127,10 +127,15 @@ pub(super) struct ConnectState {
 
     /// a context to keep track of the autoplay context
     autoplay_context: Option<StateContext>,
+
+    /// The volume adjustment per step when handling individual volume adjustments.
+    pub volume_step_size: u16,
 }
 
 impl ConnectState {
     pub fn new(cfg: ConnectConfig, session: &Session) -> Self {
+        let volume_step_size = u16::MAX.checked_div(cfg.volume_steps).unwrap_or(1024);
+
         let device_info = DeviceInfo {
             can_play: true,
             volume: cfg.initial_volume.into(),
@@ -195,6 +200,7 @@ impl ConnectState {
                 }),
                 ..Default::default()
             },
+            volume_step_size,
             ..Default::default()
         };
         state.reset();